# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Health Triage Assistant API",
    "description": "### Offline-First Health Triage Assistant API\n\nProduction-grade RESTful API serving clinical triage evaluation, emergency dispatch, health profile management, and PWA outbox sync.\n\n**Key Architectural Highlights:**\n- **Dual Intelligence**: Deterministic rule engine + optional Gemini AI enhancement\n- **Multilingual**: BCP 47 language tagging (English `en`, Twi `tw`)\n- **Clean Architecture**: Isolated domain models, services, and interface adapters\n- **RFC 7807 Errors**: Structured problem details error payloads\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/users/me/profile": {
      "get": {
        "tags": [
          "User Profile"
        ],
        "summary": "Get current user health profile",
        "description": "Returns the health profile for the currently authenticated user.",
        "operationId": "get_my_profile_api_v1_users_me_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthProfileResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "User Profile"
        ],
        "summary": "Create or update current user health profile",
        "description": "Creates or updates the health profile for the currently authenticated user.",
        "operationId": "upsert_my_profile_api_v1_users_me_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HealthProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/users/me/emergency-contacts": {
      "get": {
        "tags": [
          "User Profile"
        ],
        "summary": "Get emergency contacts",
        "description": "Returns all emergency contacts registered for the authenticated user.",
        "operationId": "get_emergency_contacts_api_v1_users_me_emergency_contacts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EmergencyContactResponse"
                  },
                  "type": "array",
                  "title": "Response Get Emergency Contacts Api V1 Users Me Emergency Contacts Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "User Profile"
        ],
        "summary": "Add or update an emergency contact",
        "description": "Creates or updates an emergency contact for the authenticated user.",
        "operationId": "upsert_emergency_contact_api_v1_users_me_emergency_contacts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmergencyContactRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyContactResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmergencyContactRequest": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "contact_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Contact Name"
          },
          "phone_number": {
            "type": "string",
            "maxLength": 20,
            "minLength": 7,
            "title": "Phone Number"
          },
          "relationship_type": {
            "type": "string",
            "maxLength": 50,
            "title": "Relationship Type",
            "default": "Other"
          },
          "is_primary": {
            "type": "boolean",
            "title": "Is Primary",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "contact_name",
          "phone_number"
        ],
        "title": "EmergencyContactRequest",
        "description": "Emergency contact create/update request."
      },
      "EmergencyContactResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "contact_name": {
            "type": "string",
            "title": "Contact Name"
          },
          "phone_number": {
            "type": "string",
            "title": "Phone Number"
          },
          "relationship_type": {
            "type": "string",
            "title": "Relationship Type"
          },
          "is_primary": {
            "type": "boolean",
            "title": "Is Primary"
          }
        },
        "type": "object",
        "required": [
          "id",
          "contact_name",
          "phone_number",
          "relationship_type",
          "is_primary"
        ],
        "title": "EmergencyContactResponse",
        "description": "Emergency contact read response."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HealthProfileRequest": {
        "properties": {
          "full_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Full Name"
          },
          "age": {
            "type": "integer",
            "maximum": 120.0,
            "minimum": 0.0,
            "title": "Age"
          },
          "biological_sex": {
            "type": "string",
            "pattern": "^(MALE|FEMALE|OTHER)$",
            "title": "Biological Sex"
          },
          "blood_group": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(A|B|AB|O)[+-]$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blood Group",
            "description": "ABO blood group with Rh factor (e.g. A+, O-)."
          },
          "chronic_conditions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Chronic Conditions"
          },
          "known_allergies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Known Allergies"
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "age",
          "biological_sex"
        ],
        "title": "HealthProfileRequest",
        "description": "Health profile create/update request payload."
      },
      "HealthProfileResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "age": {
            "type": "integer",
            "title": "Age"
          },
          "biological_sex": {
            "type": "string",
            "title": "Biological Sex"
          },
          "blood_group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blood Group"
          },
          "chronic_conditions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Chronic Conditions"
          },
          "known_allergies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Known Allergies"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "emergency_contacts": {
            "items": {
              "$ref": "#/components/schemas/EmergencyContactResponse"
            },
            "type": "array",
            "title": "Emergency Contacts"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "full_name",
          "age",
          "biological_sex",
          "blood_group",
          "chronic_conditions",
          "known_allergies",
          "updated_at"
        ],
        "title": "HealthProfileResponse",
        "description": "Health profile read response payload."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}